move get_xblock
